+++ tests/_boundary_anchor_unit_test.py
@@ -25,7 +25,7 @@ sys.modules[PACKAGE] = package
 
 shared_nodes = types.ModuleType(PACKAGE + ".nodes")
 shared_nodes.MiniMaxH3MotionContext = object
-shared_nodes._claim_inline_patch_ownership = lambda: "test patch owner"
+shared_nodes._claim_inline_patch_ownership = lambda _conditioning=None: "test patch owner"
 shared_nodes._prepare_native_guide_conditioning = lambda value: value
 shared_nodes._resize = lambda *args: None